llmspell_agents::composition::traits

Trait Composable

pub trait Composable:
    BaseAgent
    + Send
    + Sync {
    // Required methods
    fn composition_metadata(&self) -> CompositionMetadata;
    fn can_compose_with(&self, other: &dyn Composable) -> bool;
    fn exposed_capabilities(&self) -> Vec<Capability>;
    fn required_capabilities(&self) -> Vec<Capability>;
}

Trait for components that can be composed into higher-level structures

Required Methods

fn composition_metadata(&self) -> CompositionMetadata

Get the composition metadata for this component

fn can_compose_with(&self, other: &dyn Composable) -> bool

Check if this component can be composed with another

fn exposed_capabilities(&self) -> Vec<Capability>

Get the exposed capabilities of this component

fn required_capabilities(&self) -> Vec<Capability>

Get the required capabilities this component needs