llmspell_agents::monitoring::tracing

Trait TraceExporter

pub trait TraceExporter: Send + Sync {
    // Required method
    fn export(&self, span: &TraceSpan) -> Result<()>;
}

Trait for exporting traces

Required Methods

fn export(&self, span: &TraceSpan) -> Result<()>

Export a completed span

Errors

Returns an error if the span cannot be exported due to network issues, serialization failures, or backend service unavailability.